requirements.txt
fastapi
uvicorn
transformers
torch
numpy
faiss-cpu
tqdm
pydantic
gdown
